+++ packages/@tinacms/form-builder/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.14-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/form-builder@0.2.13...@tinacms/form-builder@0.2.14-alpha.0) (2020-02-06)
+
+**Note:** Version bump only for package @tinacms/form-builder
+
+
+
+
+
 ## [0.2.13](https://github.com/tinacms/tinacms/compare/@tinacms/form-builder@0.2.13-alpha.0...@tinacms/form-builder@0.2.13) (2020-01-27)