+++ packages/next-tinacms-markdown/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1-alpha.0](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.0...next-tinacms-markdown@0.2.1-alpha.0) (2020-02-06)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
 # [0.2.0](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.0-alpha.0...next-tinacms-markdown@0.2.0) (2020-02-03)
 
 **Note:** Version bump only for package next-tinacms-markdown